# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Bulk Purchasing Platform API
  description: |
    Complete API documentation for the Save Together bulk purchasing application.
    
    ## Features
    - User authentication with JWT tokens
    - Product catalog management
    - Vendor management and dashboards
    - Bulk purchasing groups
    - Order management with MOQ validation
    - Escrow-protected transactions
    - Dispute resolution
    - Real-time chat with WebSocket (Socket.IO)
    
    ## Authentication
    Most endpoints require a JWT token obtained from `/auth/login`. Include the token in the Authorization header:
    ```
    Authorization: Bearer {token}
    ```
  version: 1.0.0
  contact:
    name: API Support
    email: support@savetogether.com
  license:
    name: MIT
servers:
  - url: http://localhost:3001/api
    description: Development Server
  - url: https://api.savetogether.com
    description: Production Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from login endpoint. Format is "Bearer {token}"
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [superUser, admin, vendor, member]
          description: User's global role in the system
        avatar:
          type: string
          format: uri
          nullable: true
        vendor_id:
          type: string
          format: uuid
          nullable: true
          description: ID of vendor if role is 'vendor'
        trust_score:
          type: integer
          nullable: true
        is_active:
          type: boolean
        is_online:
          type: boolean
        created_at:
          type: string
          format: date-time
      required: [id, email, name, role, created_at]
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
      required: [success, error]
    ValidationErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: Authenticate user and receive JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  minLength: 6
                  example: password123
              required: [email, password]
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  token:
                    type: string
                    description: JWT token for API authentication
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Validation error or invalid email format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Invalid credentials or account deactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/signup:
    post:
      tags: [Authentication]
      summary: User registration
      description: Create new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                  minLength: 6
                name:
                  type: string
                  minLength: 2
                role:
                  type: string
                  enum: [member, vendor]
                  default: member
                  description: Only 'member' or 'vendor' allowed on signup
              required: [email, password, name]
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Email exists or validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh JWT token
      security:
        - bearerAuth: []
      responses:
        '200':
          description: New token generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  token:
                    type: string

  # ============ USERS ============
...